+++ libraries/lps/include/mcrl2/lps/disjointness_checker.h
@@ -32,10 +32,10 @@
 class Disjointness_Checker {
   private:
     /// \brief The number of summands of the LPS passed as argument of the constructor.
-    int f_number_of_summands;
+    size_t f_number_of_summands;
 
     /// \brief The number of parameters of the LPS passed as argument of the constructor.
-    int f_number_of_parameters;
+    size_t f_number_of_parameters;
 
     /// \brief An indexed set of parameters.
     ATermIndexedSet f_parameter_set;
@@ -47,14 +47,14 @@ class Disjointness_Checker {
     bool* f_changed_parameters_per_summand;
 
     /// \brief Updates the array Disjointness_Checker::f_used_parameters_per_summand, given the expression a_expression.
-    void process_data_expression(int a_summand_number, ATermAppl a_expression);
+    void process_data_expression(size_t a_summand_number, ATermAppl a_expression);
 
     /// \brief Updates the array Disjointness_Checker::f_used_parameters_per_summand, given the multiaction a_multi_action.
-    void process_multi_action(int a_summand_number, ATermAppl a_multi_action);
+    void process_multi_action(size_t a_summand_number, ATermAppl a_multi_action);
 
     /// \brief Updates the arrays Disjointness_Checker::f_changed_parameters_per_summand and
     /// \brief Disjointness_Checker::f_used_parameters_per_summand, given the summand a_summand.
-    void process_summand(int a_summand_number, ATermAppl a_summand);
+    void process_summand(size_t a_summand_number, ATermAppl a_summand);
 
     /// \brief Initializes the indexed set Disjointness_Checker::f_parameter_set, given the list of parameters a_parameters.
     void process_parameters(ATermList a_parameters);
@@ -74,7 +74,7 @@ class Disjointness_Checker {
     ~Disjointness_Checker();
 
     /// \brief Indicates whether or not the summands with number a_summand_number_1 and a_summand_number_2 are disjoint.
-    bool disjoint(int a_summand_number_1, int a_summand_number_2);
+    bool disjoint(size_t a_summand_number_1, size_t a_summand_number_2);
 };
 
 #endif
